% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_BCH.R
\name{mx_BCH}
\alias{mx_BCH}
\title{mx_BCH}
\usage{
mx_BCH(
  formula.tmb = NULL,
  data = NULL,
  post.prob = NULL,
  prior.prob = NULL,
  family = gaussian(),
  robust.se = "none",
  id = NULL,
  reference_group = 1
)
}
\arguments{
\item{formula.tmb}{Specify the formula object: "latent_class"}

\item{data}{Your data}

\item{post.prob}{A list of latent class probability columns}

\item{prior.prob}{Prior probabilities; if NULL, will be calculated}

\item{family}{A GLM object specifying the error distribution. "Gaussian" assumes normal error distribution}

\item{robust.se}{Robust standard error correction. Currently set to "none", as it is not available}

\item{id}{A column in data that has original ID numbers}

\item{reference_group}{The reference group when latent class is the predictor}
}
\value{
A list of class \code{"mxGlm"} containing:
\describe{
  \item{\code{model}}{The fitted model object returned by \code{mxBCHfit}.}
  \item{\code{data}}{A long-format data frame with class assignment and BCH weights.}
  \item{\code{formula}}{The final formula used in model estimation.}
}
}
\description{
This function uses BCH to predict distal outcomes by latent class membership in a GLM framework.
}
